pyproject.toml
[project]
name = "concurrency-bench"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.42.2",
    "openhands-sdk>=1.3.0",
    "openhands-tools>=1.3.0",
    "setuptools>=80.9.0",
]

[build-system]
requires = ["uv_build>=0.9.0,<0.10.0"]
build-backend = "uv_build"
